requirements.txt
langchain==0.1.10
langgraph==0.1.3
langchain-community==0.0.27
langchain-google-vertexai==0.0.1
langsmith==0.0.97
chromadb==0.4.22
google-cloud-aiplatform==1.39.0
sentence-transformers==2.5.0
tiktoken==0.5.2
streamlit==1.32.2
pymupdf==1.23.22
python-dotenv==1.0.0
unstructured==0.12.5
tavily-python==0.3.1
requests==2.31.0